requirements.txt
pyyaml==5.4.1
psutil==5.9.0
scapy==2.5.0
